Add a callback function for page naming
This allows you to pass any number of things to use instead of a page
number. In my case I have a list of ~160 faculty that needs to be easy
to navigate. Numbering the pages doesn't make as much sense as splitting
it like so: A-C, C-F, G-H, etc. This function allows you to do more
complex changes like that, or even simpler ones like mapping language
translations.